Code refactoring by using repositories for statistics.

+++ src/main/java/com/pepe/albarapp/service/dto/statistics/RankingDto.java
@@ -1,7 +1,5 @@
 package com.pepe.albarapp.service.dto.statistics;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,6 +10,6 @@ import lombok.NoArgsConstructor;
 public class RankingDto {
 
 	private String customerAlias;
-    private BigDecimal invoiceTotal;
+    private Double invoiceTotal;
     
 }
